+++ core/src/test/scala/org/apache/spark/sql/catalyst/plans/BasePlanTest.scala
@@ -159,24 +159,5 @@ class BasePlanTest extends BaseTiSparkTest {
     assert(estimatedRowCount === answer)
   }
 
-  test("baseline test") {
-    val tableName = "full_data_type_table_idx"
-    val df = spark.sql(
-      s"select tp_bigint from $tableName where tp_tinyint = 0 and tp_int < 0"
-    )
 
-    checkIndex(df, "idx_tp_tinyint_tp_int")
-
-    assertThrows[TestFailedException] {
-      checkIsTableScan(df, tableName)
-    }
-
-    checkIsIndexScan(df, tableName)
-
-    assertThrows[TestFailedException] {
-      checkIsCoveringIndexScan(df, tableName)
-    }
-
-    checkEstimatedRowCount(df, tableName, 2)
-  }
 }
